Update submiterator.py
small bug fixes

+++ submiterator.py
@@ -1,3 +1,5 @@
+import os
+
 settings = open("README", 'r')
 lines = settings.readlines()
 settings.close()
@@ -21,6 +23,9 @@ for line in lines:
                 value = value[:-1]
             dict[x[0]] = value
 
+if not os.path.exists(dict["locationofCLT"]) or dict["locationofCLT"][-1] == '/':
+    raise Exception("Error: check the 'locationofCLT' specification in your README file for errors.")
+
 if dict["rewriteProperties"] == "yes":
     old_properties_file = open(dict["locationofCLT"] + "/bin/mturk.properties", 'r').readlines()
     backup = open(dict["locationofCLT"] + "/bin/mturk.properties.backup", 'w')
@@ -31,17 +36,17 @@ if dict["rewriteProperties"] == "yes":
     if (dict["liveHIT"] == "yes"):
         for line in old_properties_file:
             if "://mechanicalturk.sandbox.amazonaws.com/?Service=AWSMechanicalTurkRequester" in line:
-                new_properties_file.write("# service_url=https://mechanicalturk.sandbox.amazonaws.com/?Service=AWSMechanicalTurkRequester")
+                new_properties_file.write("# service_url=https://mechanicalturk.sandbox.amazonaws.com/?Service=AWSMechanicalTurkRequester\n")
             elif "://mechanicalturk.amazonaws.com/?Service=AWSMechanicalTurkRequester" in line:
-                 new_properties_file.write("service_url=https://mechanicalturk.amazonaws.com/?Service=AWSMechanicalTurkRequester")
+                 new_properties_file.write("service_url=https://mechanicalturk.amazonaws.com/?Service=AWSMechanicalTurkRequester\n")
             else:
                 new_properties_file.write(line)
     else:
         for line in old_properties_file:
             if "://mechanicalturk.sandbox.amazonaws.com/?Service=AWSMechanicalTurkRequester" in line:
-                new_properties_file.write("service_url=https://mechanicalturk.sandbox.amazonaws.com/?Service=AWSMechanicalTurkRequester")
+                new_properties_file.write("service_url=https://mechanicalturk.sandbox.amazonaws.com/?Service=AWSMechanicalTurkRequester\n")
             elif "://mechanicalturk.amazonaws.com/?Service=AWSMechanicalTurkRequester" in line:
-                new_properties_file.write("# service_url=https://mechanicalturk.amazonaws.com/?Service=AWSMechanicalTurkRequester")
+                new_properties_file.write("# service_url=https://mechanicalturk.amazonaws.com/?Service=AWSMechanicalTurkRequester\n")
             else:
                 new_properties_file.write(line)
     new_properties_file.close()
